pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pink-telegram-transcriber"
version = "1.0.0"
description = "Simple Telegram bot for voice transcription using Pink Transcriber service"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "pinkhairedboy", email = "pink@pinkhaired.com"}
]
keywords = ["telegram", "transcription", "voice-to-text", "bot"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
]

dependencies = [
    "python-telegram-bot==22.5",
    "python-dotenv==1.0.0",
    "psutil>=6.1.0",
]

[project.urls]
Homepage = "https://github.com/pinkhairedboy/pink-telegram-transcriber"
Repository = "https://github.com/pinkhairedboy/pink-telegram-transcriber"

[project.scripts]
pink-telegram-transcriber = "pink_telegram_transcriber.bot:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
